Moved code that validates coordinates of the selected feature and then updates it in "editing" array into separate function.
This function is now running in two epics that needs "editing" array to be updated with valid geoJson of selected geometry.
